requirements.txt
Django==5.2.8
djangorestframework==3.15.2
django-filter==24.2
django-cors-headers==4.4.0
mysqlclient==2.2.4
python-dotenv==1.0.1
cryptography==43.0.1
PyMySQL==1.1.1

